init: start of usernetes in python
We need to be able to orchestrate user space kubernetes
easily, and alongside flux. Our current approaches use
the GitHub repository and a Makefile, and that works
well (and possibly we should stick with it) but if we
want to integrate usernetes with flux via a script,
our best bet is the Python SDK, and so I have chosen
this approach. It might not be the best idea, and it
still is not clear what the deployment strategy will
be (e.g., clone release on the fly, to home, or copy
from the install) but we will figure this out with
further testing. The general goal is to streamline
the deployment and integrate tightly with flux.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>